Fix Windows build where checking UTS namespace doesn’t make sense
